+++ snitch/__init__.py
@@ -20,7 +20,7 @@ __all__ = [
     "get_notification_model",
 ]
 
-__version__ = "2.2.1"
+__version__ = "2.3.0"
 
 
 def autodiscover():