+++ tests/settings.py
@@ -32,7 +32,7 @@ TEMPLATES = [
         },
     }
 ]
-DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
+DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
 
 # DJANGO PUSH NOTIFICATIONS
 # ------------------------------------------------------------------------------